:recycle: api: Refacto getNextChallengeForCertification to split it for v2/v3
