+++ api/src/certification/course/domain/usecases/get-next-challenge-for-v2-certification.js
@@ -1,0 +1,13 @@
+const getNextChallengeForV2Certification = async function ({
+  assessment,
+  certificationChallengeRepository,
+  challengeRepository,
+}) {
+  return certificationChallengeRepository
+    .getNextNonAnsweredChallengeByCourseId(assessment.id, assessment.certificationCourseId)
+    .then((certificationChallenge) => {
+      return challengeRepository.get(certificationChallenge.challengeId);
+    });
+};
+
+export { getNextChallengeForV2Certification };
