+++ api/src/certification/course/domain/usecases/index.js
@@ -42,7 +42,8 @@ const usecasesWithoutInjectedDependencies = {
       'find-certification-attestations-for-division.js',
       'get-certification-attestation.js',
       'get-certification-attestations-for-session.js',
-      'get-next-challenge-for-certification.js',
+      'get-next-challenge-for-v2-certification.js',
+      'get-next-challenge-for-v3-certification.js',
       'get-session-certification-reports.js',
       'get-v3-certification-course-details-for-administration.js',
       'reject-certification-course.js',
